Create Method_Overloading.cs
Function or Method overloading allows a class to have multiple methods with the same name but with a different signature. Methods can be overloaded based on the number, type (int, float, etc.), and kind (value, ref, out) of parameters. Signature of a method consists of name, type, kind and number of parameters.