Create Class_vs_Struct
Structures are value types like integers, floats, etc and are stored on the stack whereas classes, which are reference types, are stored on the heap. You can copy a struct but not a class. Structure cannot have explicit parameter-less constructors or destructors however class can have both. Structures cannot inherit from another class but a class can. Both structure and class can inherit from an interface. A class or struct cannot inherit from another struct because a struct is sealed.