Create Abstract_Classes
The abstract keyword is used to create abstract classes. An abstract class is incomplete and hence cannot be instantiated. An abstract class can only be used as base class and cannot be sealed. An abstract class any contain abstract members (methods, properties, indexers, and events) but it's not mandatory. A non-abstract class derived from an abstract class must provide implementations for all inherited abstract members, if not then make the class as abstract.