Create Delegates
A delegate is a type-safe function pointer which holds a reference (pointer) to the function. The signature of the delegate must match the signature of the function that it points to or we'll get a compiler error - this is why delegates are called type-safe. A delegate, similar to a class, you can create an instance of it and pass in the function name to which it points to as a parameter to the delegate constructor. Delegates look very similar to a method with the delegate keyword.