Rename Delegates to Delegates.cs
Delegate is a type safe function pointer that holds a reference (pointer) to a function. The signature of the delegate must match the signature of the function that it points to or you'll get a compiler error. This is why they're called a type safe function pointer. A delegate is similar to a class because you can create an instance of it and pass in the function name as a parameter which is where the delegate points. Delegate syntax is very similar to a method with the delegate keyword.