Update HowWhereIndexersUsed.cs
These two files show how and where indexers are used. The first file stores data in a session-state variable and retrieves the data as either an index or as an integral indexer or a string indexer. The second file retrieves data from a SQL table and writes it to a Web Form. using ADO.NET Core.