Create MethodParametersOptionalParameterDefaults.cs
To make method parameters optional one of the four ways is by specifying parameter defaults. In the top example we set the method's third parameter to null. In the bottom example we use Name Parameters to specify which parameter we want to apply the argument to when we call the method.