Create ListCollections.cs
List of on of the generic collection classes present in System.Collection.Generic namespace. A List class can be used to create a collection of any type like integers, strings and complex types. The object stored in the List can be access by index. Unlike arrays, lists can grow in size automatically