Create SortListComplexTypes.cs
Sorting a list of complex types uses IComparable interface and provide implementation for CompareTo method. If you prefer to not use the Sort functionality then you can use provide your own by implementing IComparer interface then pass an instance of the class that implements IComparer interface as an argument to the Sort method.