Create SortListComplexTypesComparisonDelegate.cs
When using Comparison delegate with List first create a function whose signature matches the Comparison delegate, then create an instance of the Comparison delegate which points to the function (by passing it the function's name), then pass that instance as an argument to the Sort method. There are two easier approaches to accomplish this by using the delegate keyword or lambda expression.