Create GenericStackCollectionClass.cs
Stack is a generic LIFO collection class. To insert an item at the top of the stack use Push method. To remove and return an item at the top of the stack use the Pop method. Use Peek method to return the items at the top of the stack without removing it. Use a foreach loop to iterate through the items but it won't remove them. Use the Contains method to check if an item exists in the stack.