Update ParameterizedThreadStartDelegate.cs
We use ParameterizedThreadStart over ThreadStart delegate is we have some data that we want to pass down to the Thread function. Keep in mind that using ParameterizedThreadStart and Thread.Start(Object) method to pass data to the Thread function is not type safe because they operate on the object datatype therefore any data can be passed down. ParameterizedThreadStart should be avoided.