Create PassingDataThreadCallbackMethod.cs
We can retrieve data from Thread function using Callback method. First we retrieve the target number from user in the Main thread then the Main thread should create a child thread and pass that target number to the child thread. Then the child thread will compute the sum of number and return the sum to the Main thread using a Callback function. Finally the Callback function will print the sum. The first step is to create a callback delegate.